requirements.txt
coverage>=4.5.4
Cython>=0.29.16
matplotlib>=3.0.3
tensorflow>=2.3.0
tensorflow_probability>=0.9.0
tensorboard>=2.2.0
tqdm>=4.31.1
tensorflow_addons>=0.9.0
numpy>=1.18,<1.19.0
Pillow>=6.2.1
protobuf>=3.10.0
pylint
PyYAML>=5.1.2
keras-autodoc==0.4.0
mkdocs>=1.0.4
mkdocs-material>=4.6.3
Markdown>=3.2.1
pycocotools;platform_system!="Windows"
pycocotools-windows;platform_system=="Windows"
